@@ -1,6 +1,6 @@
 {
   "name": "choerodon-front-boot",
-  "version": "0.6.5",
+  "version": "0.7.0",
   "description": "Tools to build website for Choerodon",
   "main": "lib/containers/components/index.js",
   "bin": {
@@ -51,9 +51,10 @@
     "case-sensitive-paths-webpack-plugin": "^2.1.2",
     "chalk": "^2.4.1",
     "change-case": "^3.0.1",
-    "choerodon-ui": "0.3.4",
+    "choerodon-ui": "0.3.5",
     "classnames": "^2.2.5",
     "commander": "^2.9.0",
+    "component-classes": "^1.x",
     "css-loader": "^0.25.0",
     "eslint": "^4.2.0",
     "eslint-config-airbnb": "^15.1.0",
@@ -64,6 +65,7 @@
     "extract-text-webpack-plugin": "^3.0.0",
     "file-loader": "^0.9.0",
     "friendly-errors-webpack-plugin": "^1.7.0",
+    "glob": "^7.1.2",
     "history": "^4.6.3",
     "html-webpack-plugin": "^3.2.0",
     "json-loader": "^0.5.7",
@@ -81,9 +83,9 @@
     "postcss-loader": "^2.1.5",
     "prop-types": "^15.5.10",
     "query-string": "^4.1.0",
-    "react": "^16.3.1",
+    "react": "^16.4.1",
     "react-dev-utils": "^5.0.1",
-    "react-dom": "^16.3.1",
+    "react-dom": "^16.4.1",
     "react-hot-loader": "^3.1.1",
     "react-intl": "^2.3.0",
     "react-router-dom": "^4.1.1",
