@@ -1,6 +1,7 @@
-const path = require('path');
-
 module.exports = {
   routes: { test: './test/routeIndex' },
-  server: 'https://api.choerodon.com.cn',
+  server: 'http://api.staging.saas.hand-china.com',
+  dashboard: {
+    'choerodon-front-iam': 'test/dashboard/*',
+  },
 };
