@@ -1,50 +1,53 @@
 /*eslint-disable*/
 const zh_CN = {
-  "user.preferences": "用户首选项",
-  "signOut": "安全退出",
-  "choose": "请选择",
-  "select": "选择",
-  "cancel": "返回",
+  'user.preferences': '用户首选项',
+  'signOut': '安全退出',
+  'choose': '请选择',
+  'select': '选择',
+  'cancel': '返回',
 
-  "learnmore": "了解更多",
-  "boot.signout": "安全退出",
-  "boot.select": "选择",
-  "boot.return": "返回",
+  'learnmore': '了解更多',
+  'boot.signout': '安全退出',
+  'boot.select': '选择',
+  'boot.return': '返回',
 
   // operation
-  "boot.operation": "操作",
-  "boot.add": "新增",
-  "boot.create": "创建",
-  "boot.edit": "编辑",
-  "boot.modify": "修改",
-  "boot.delete": "删除",
-  "boot.remove": "移除",
-  "boot.confirm.delete": "确认删除",
-  "boot.cancel": "取消",
-  "boot.refresh": "刷新",
-  "boot.detail": "详情",
-  "boot.enable": "启用",
-  "boot.disable": "停用",
+  'boot.operation': '操作',
+  'boot.add': '新增',
+  'boot.create': '创建',
+  'boot.edit': '编辑',
+  'boot.modify': '修改',
+  'boot.delete': '删除',
+  'boot.remove': '移除',
+  'boot.confirm.delete': '确认删除',
+  'boot.cancel': '取消',
+  'boot.refresh': '刷新',
+  'boot.detail': '详情',
+  'boot.enable': '启用',
+  'boot.disable': '停用',
 
   // status
   // success
-  "boot.operation.success": "操作成功",
-  "boot.create.success": "创建成功",
-  "boot.modify.success": "修改成功",
-  "boot.delete.success": "删除成功",
-  "boot.remove.success": "移除成功",
-  "boot.enable.success": "启用成功",
-  "boot.disable.success": "停用成功",
+  'boot.operation.success': '操作成功',
+  'boot.create.success': '创建成功',
+  'boot.modify.success': '修改成功',
+  'boot.delete.success': '删除成功',
+  'boot.remove.success': '移除成功',
+  'boot.enable.success': '启用成功',
+  'boot.disable.success': '停用成功',
   // error
-  "boot.operation.error": "操作失败",
-  "boot.create.error": "创建失败",
-  "boot.modify.error": "修改失败",
-  "boot.delete.error": "删除失败",
-  "boot.remove.error": "移除失败",
-  "boot.enabled.error": "启用失败",
-  "boot.disable.error": "停用失败",
+  'boot.operation.error': '操作失败',
+  'boot.create.error': '创建失败',
+  'boot.modify.error': '修改失败',
+  'boot.delete.error': '删除失败',
+  'boot.remove.error': '移除失败',
+  'boot.enabled.error': '启用失败',
+  'boot.disable.error': '停用失败',
 
-
+  'boot.dashboard.home': '首页',
+  'boot.dashboard.customize': '自定义',
+  'boot.dashboard.complete': '完成',
+  'boot.dashboard.success': '保存成功',
 };
 
 export default zh_CN;
