@@ -10,7 +10,7 @@ function getDefaultLanguage() {
 }
 
 class AppState {
-  @observable menuType = {}; // 一个菜单对象 {id:'',name:'',type:''}
+  @observable menuType = null; // 一个菜单对象 {id:'',name:'',type:''}
   @observable expanded = false;
   @observable userInfo = {};
   @observable debugger = false; // 调试模式
