@@ -2,10 +2,17 @@ import AppState from './AppState';
 import HeaderStore from './HeaderStore';
 import MenuStore from './MenuStore';
 import PermissionStore from './PermissionStore';
+import DashboardStore from './DashboardStore';
+import { dashboard } from '../common';
 
-export default {
+const stores = {
   AppState,
   HeaderStore,
   MenuStore,
   PermissionStore,
 };
+if (dashboard) {
+  stores.DashboardStore = DashboardStore;
+}
+
+export default stores;
